// server/modules/workout/seed-data/body-target.enum.ts
export enum ENUM_SYSTEM_BODY_TARGET {
  CHEST = "chest",
  SHOULDERS_ANTERIOR = "shoulders_anterior",
  SHOULDERS_LATERAL = "shoulders_lateral",
  SHOULDERS_POSTERIOR = "shoulders_posterior",
  TRICEPS = "triceps",
  BICEPS = "biceps",
  FOREARMS = "forearms",
  LATS = "lats",
  UPPER_BACK = "upper_back",
  MID_BACK = "mid_back",
  LOWER_BACK = "lower_back",
  TRAPS = "traps",
  ERECTOR_SPINAE = "erector_spinae",
  CORE_ABS = "core_abs",
  CORE_DEEP = "core_deep",
  OBLIQUES = "obliques",
  HIP_FLEXORS = "hip_flexors",
  GLUTES = "glutes",
  HAMSTRINGS = "hamstrings",
  QUADRICEPS = "quadriceps",
  ADDUCTORS = "adductors",
  CALVES = "calves",
  CARDIOVASCULAR = "cardiovascular",
  FLEXIBILITY = "flexibility",
  BALANCE = "balance",
  ENDURANCE = "endurance",
}